Move test objects to a separate file.
